# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Match Day",
    "description": "Microsserviços para aluguel de Quadras Esportivas"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
    "/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "List user by id",
        "description": "This endpoint will list the user account by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new user",
        "description": "This endpoint will create a new customer account",
        "responses": {
          "201": {
            "description": "Cliente cadastrado com sucesso"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sportCourt/{id}": {
      "get": {
        "tags": [
          "Court"
        ],
        "summary": "List court by id",
        "description": "This endpoint will list the courts by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/sportCourt": {
      "post": {
        "tags": [
          "Court"
        ],
        "summary": "Create a court",
        "description": "Endpoint to create a court",
        "parameters": [
          {
            "name": "CourtPayload",
            "in": "body",
            "description": "Payload",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Nome da Quadra"
                },
                "reservated": {
                  "type": "boolean",
                  "example": true
                },
                "description": {
                  "type": "string",
                  "example": "Quadra de futebol"
                },
                "size": {
                  "type": "string",
                  "example": "5"
                },
                "location": {
                  "type": "string",
                  "example": "Ijui"
                },
                "start_datetime": {
                  "type": "string",
                  "example": "2024-06-12T12:00:00Z"
                },
                "end_datetime": {
                  "type": "string",
                  "example": "2024-06-12T14:00:00Z"
                },
                "price": {
                  "type": "number",
                  "example": 100
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "reservated": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "size": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "start_datetime": {
                    "example": "any"
                  },
                  "end_datetime": {
                    "example": "any"
                  },
                  "price": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Court"
        ],
        "summary": "List all courts",
        "description": "This endpoint will list all the courts",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}